Add items relating to call for speakers. See vtcodecamp/documents#4.
Since the Diversity Statement was originally published after the call for session proposals closed for 2016, it was determined to publish these list items after the 2016 event since it was effectively too late for people to act on them for the 2016 event. It felt more sincere to publish them at a time when they were actionable by people for the upcoming year (2017).